refactor(form): restructure directory and create custom form field affter reviews
