+++ src/app/todos/page.tsx
@@ -1,0 +1,7 @@
+import TodoList from '@/src/section/Todos';
+
+const TodoPage = () => {
+  return <TodoList />;
+};
+
+export default TodoPage;
