+++ src/section/Todos/index.tsx
@@ -1,0 +1,5 @@
+const TodoList = () => {
+  return <div>Todo List</div>;
+};
+
+export default TodoList;
